requirements.txt
mcrcon~=0.6.0
python-telegram-bot~=12.5.1
